move function ref to a separate file
